Update code to show subdirectories
The new code behaves better but does not implement a true recursive clickable directory tree. See the ls capability in the SD library which does iterate over all files because this will not recurse to further depth than one. Cleaver coders who wish to fiddle with this are welcome to get the code and submit a request with your edits.