Changes to remove temp global variable
Per https://3.basecamp.com/3732686/buckets/4383362/todos/1319808931#__recording_1402164150

John, I took the check to see if the acceleration was positive and put that check directly in the if statement. Then the code to set tilt is not needed and the variable declaration is unneeded. Please review & try and then if this PR or your edits are good, we can look to merge.