Update startup colors to be more primary
Star Trek was one of the first color television shows so they had a very primary palette of colors. Changed intro swirl to RED, GREEN, BLUE as Limor asked that the pastel colors be more Star Trek.

Dano should test this - if it isn't enough Trek, maybe a different animation?