removed micropython-specific code
The update to this PR removes `micropython`-specific code. Tested on an adafruit feather board with the following results:

```
Computing the RMS value of 100 numbers
traditional          :    5.472ms [result=3535.843611]
ulab, with ndarray, some implementation in python :    0.500ms [result=3535.853624]
ulab only, with list :    0.655ms [result=3535.854340]
ulab only, with ndarray :    0.139ms [result=3535.854340]
```
The result of the second run might be misleading (seems to be faster than the third), because it does not contain the costs of the list-to-ndarray conversion (since that happens earlier, outside of time measurement).