Adding code and files for Circuit Playground Star Locket learn guide
Adding the two bitmaps, audio file (arranged and recorded for robot xylophone) and code. Uses a Circuit Playground Bluefruit and TFT Gizmo. Default behavior is for the crescent moon to rotate around the screen. If you press the A button you can turn on or off the music. If you press the B button you can pause or resume the animation.