Add a workflow to generate folder images
This workflow will generate a bunch of images and place them on a branch called
"folder-images", which (once enabled via github pages) will appear on
adafruit.github.io.

The workflow is triggered:
 * on commits to the main branch, including merged pull requests
 * daily, after the bundle is updated
 * on request via the github web page

Typically, this means that the images will be available (well) within an hour
after a pull request is merged, and will be updated daily if required by
changes to the requirements of libraries.