Fix parsing the vaccination data
The code apparently broke when a new column was added to the CSV
data to show the number of boosters.  This field is blank in the US
data, so the code would crash trying to turn the empty string into an
integer.

Since upstream seems to add columns at the far right (a sensible
choice) we should index using positive numbers instead.

This also changes the code so that in the event of a failure it actually
*DOES* retry. I designed it to wait for 1 hour in deep sleep before retrying,
so that it doesn't immediately run down the battery in case of transient
errors.