Update local color mid-grey to RGB565
A hardcoded value of 0x808080 was in the code for a mid-level grey. This is an RGB888 value, but the ST77XX library uses RGB565 values. A mid-grey in RGB565 is 0x8410. A local define was done to define this as STS77XX_GREY and the new value placed in the calls previously using 0x808080. This solves a compiler integer overflow warning preventing CI from completing.