Add code comment for Accel Range
Accelerometer range is basically hit detection sensitivity.  Had a beginner in Discord that didn't know how to change it because it's not documented in the learn guide.  Need a better way to make this parameter configurable and easy to use. Please also add something into the learn guide about using this for lowering sensitivity.  By default the lightsaber is set to 2G which is its most sensitive setting.