pi-hole / bookworm / st7789 changes
A lot has changed since this guide was release.

The Pi Hole 6.x release in early 2025 dramatically changed the API.

Bookworm changes how some of the pins are used.

Even the ST7789 MiniPiTFT chip has some different options.

This code was tested and work with a MiniPiTFt on a Pi 4. I'd like to update the guide to support accommidate the software updates.